feat: restructure CLI tests
