+++ .gitignore
@@ -7,3 +7,5 @@ bin
 
 # Workflow
 junit.xml
+
+nul
