--- src/tests/cli-tests/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "cli-tests",
-  "version": "1.0.0",
-  "title": "zksync-cli CLI tests",
-  "description": "Auto tests for verifying zksync-cli CLI",
-  "repository": "https://github.com/matter-labs/zksync-cli",
-  "main": "index.js",
-  "private": true,
-  "scripts": {
-    "test": "npx jest --verbose --testPathPattern="
-  },
-  "keywords": [],
-  "author": "Matter Labs",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/jest": "^29.5.11",
-    "@types/shelljs": "^0.8.15",
-    "jest": "^29.7.0",
-    "shelljs": "^0.8.5",
-    "ts-jest": "^29.1.2",
-    "typescript": "^5.3.3"
-  },
-  "dependencies": {
-    "dotenv": "^16.4.1"
-  }
-}
