+++ test/helper.ts
@@ -1,0 +1,14 @@
+import * as shell from "shelljs";
+
+export interface CommandResult {
+  output: string;
+  exitCode: number;
+}
+
+export const executeCommand = (command: string): CommandResult => {
+  const result = shell.exec(command, { async: false });
+  return {
+    exitCode: result.code,
+    output: result.stdout.trim() || result.stderr.trim(),
+  };
+};
